Rename front end GraphQL files
