+++ assets/js/components/Collection/CollectionImageModal.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useMutation } from "@apollo/react-hooks";
-import { SET_COLLECTION_IMAGE } from "./collection.query";
+import { SET_COLLECTION_IMAGE } from "./collection.gql";
 import { toastWrapper } from "../../services/helpers";
 import UIFormInput from "../UI/Form/Input";
