+++ assets/js/components/Collection/collection.gql.js
