+++ assets/js/components/IngestSheet/Upload.jsx
@@ -4,9 +4,9 @@ import PropTypes from "prop-types";
 import Error from "../UI/Error";
 import Loading from "../UI/Loading";
 import { useHistory } from "react-router-dom";
-import { GET_PROJECT } from "../Project/project.query";
+import { GET_PROJECT } from "../Project/project.gql.js";
 import { useMutation } from "@apollo/react-hooks";
-import { CREATE_INGEST_SHEET } from "./ingestSheet.query";
+import { CREATE_INGEST_SHEET } from "./ingestSheet.gql.js";
 import { toastWrapper } from "../../services/helpers";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useForm } from "react-hook-form";
