+++ assets/js/components/Work/Tabs/Administrative.jsx
@@ -3,14 +3,14 @@ import PropTypes from "prop-types";
 import { useQuery, useMutation } from "@apollo/react-hooks";
 import useIsEditing from "../../../hooks/useIsEditing";
 import { toastWrapper } from "../../../services/helpers";
-import { GET_COLLECTIONS } from "../../Collection/collection.query";
-import { UPDATE_WORK, GET_WORK } from "../work.query";
+import { GET_COLLECTIONS } from "../../Collection/collection.gql.js";
+import { UPDATE_WORK, GET_WORK } from "../work.gql.js";
 import { useForm } from "react-hook-form";
 import { Link } from "react-router-dom";
 import UIFormSelect from "../../UI/Form/Select";
 import UIFormField from "../../UI/Form/Field";
 import WorkTabsHeader from "./Header";
-import { CODE_LIST_QUERY } from "../controlledVocabulary.query.js";
+import { CODE_LIST_QUERY } from "../controlledVocabulary.gql.js";
 import UICodedTermItem from "../../UI/CodedTerm/Item";
 import UIFormFieldArray from "../../UI/Form/FieldArray";
 import UIFormInput from "../../UI/Form/Input.jsx";
