+++ assets/js/screens/Project/List.test.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ScreensProjectList from "./List";
-import { GET_PROJECTS } from "../../components/Project/project.query";
+import { GET_PROJECTS } from "../../components/Project/project.gql.js";
 import { renderWithRouterApollo } from "../../services/testing-helpers";
 
 const mocks = [
